DS-1064: Extract general S3Client implementation to make it usable by Hub as well
- Extract Job service specific parts into a sub-class
- Move S3Client to xyz-util module so it can be re-used by other modules
  - Generalize related common service configuration settings into BaseConfig
- Remove unused AsyncS3Client (asynchronicity should be rather implemented on the caller side depending on the use-case)
