+++ xyz-util/src/main/java/com/here/xyz/util/service/BaseConfig.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017-2024 HERE Europe B.V.
+ * Copyright (C) 2017-2025 HERE Europe B.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,6 +20,7 @@
 package com.here.xyz.util.service;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import java.net.URI;
 import java.util.Arrays;
 import java.util.List;
 
@@ -93,6 +94,23 @@ public class BaseConfig {
    * The AWS region this service is running in. Value is <code>null</code> if not running in AWS.
    */
   public String AWS_REGION;
+
+  /**
+   * The localstack endpoint to use it for all AWS clients when running locally
+   *
+   * NOTE: This config variable may only be set when running locally, because the system uses it as an indicator
+   *  to determine whether it's running locally or on AWS.
+   */
+  public URI LOCALSTACK_ENDPOINT;
+
+  /**
+   * A comma separated list of AWS regions from which to not allow reading data (e.g. for import)
+   */
+  public String FORBIDDEN_SOURCE_REGIONS;
+
+  public List<String> forbiddenSourceRegions() {
+    return fromCommaSeparatedList(FORBIDDEN_SOURCE_REGIONS);
+  }
 
   /**
    * When set, modifies the Stream-Info header name to the value specified.
